world.c: Use calloc() instead of malloc()
It seems that malloc() clears the memory on Windows but not on
linux. Use calloc() instead to ensure memory is cleared and
call it even if zero memory is needed, this ensured a valid
pointer always to be returned.
This fixes the crash on exit when random values could be
passed to free();